requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
streamlit==1.22.0
sympy==1.12
